+++ HISTORY.md
@@ -1,6 +1,20 @@
 Work in progress
 ----------------
+- LaTeX macros / environments
+  - new package subfiles: macros \\subfile, \\subfileinclude
+    (issue [#185](../../issues/185))
+  - builtins
+    - fixed \\hspace, only add space if argument is not of zero length;
+      **thanks to @torik42** (PR [#182](../../pull/182))
+    - added macro \\paragraph(\*) (issue [#192](../../issues/192))
+  - package graphicx: added macro \\graphicspath
+    (issue [#186](../../issues/186))
+  - package tikz: now loads package graphicx (issue [#186](../../issues/186))
+- yalafi.shell: on --include, also respect \\subfile, \\subfileinclude
+- yalafi core: added file names in LaTeX-related error messages to stderr
+  (issue [#169](../../issues/169))
 - added CONTRIBUTING.md (issue [#167](../../issues/167))
+- README.md: added section on multi-file projects
 
 Version 1.3.0 (2021/01/31)
 --------------------------